2008-09-16  Bill Holmes  <billholmes54@gmail.com>
	* src/gunicode.c : Fix a warning.

	Contributed under MIT/X11 license.

svn path=/trunk/mono/; revision=113171


Commit migrated from https://github.com/mono/mono/commit/f94097b919106a99db817d22e052a6b3b3dcdbd4

+++ src/mono/eglib/src/gunicode.c
@@ -346,7 +346,7 @@ g_utf8_strlen (const gchar *str, gssize max)
 	else if (max < 0)
 		byteCount = max;
 	while (*ptr != 0 && byteCount <= max) {
-		guint cLen = g_trailingBytesForUTF8 [*ptr] + 1;
+		gssize cLen = g_trailingBytesForUTF8 [*ptr] + 1;
 		if (max > 0 && (byteCount + cLen) > max)
 			return length;
 		ptr += cLen;
